test: Add logicalSectorSize in the e2e tests

+++ pkg/azuredisk/controllerserver.go
@@ -296,7 +296,7 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 			SourceType:          sourceType,
 			DiskEncryptionSetID: diskEncryptionSetID,
 			MaxShares:           int32(maxShares),
-			LogicalSectorSize:   logicalSectorSize,
+			LogicalSectorSize:   int32(logicalSectorSize),
 		}
 		diskURI, err = d.cloud.CreateManagedDisk(volumeOptions)
 		if err != nil {
